requirements.txt
PySide6>=6.6.0
vdf>=3.4
Pillow>=10.0.0
PyInstaller>=6.0.0
